+++ include/toolbox/data/literals.h
@@ -11,6 +11,7 @@ namespace toolbox {
 namespace data {
 namespace literals {
 
+#if defined(__GNUC__) && (__GNUC__ >= 5)
 inline uint8_t operator""_byte(unsigned long long val) {
     return (uint8_t) val;
 }
@@ -23,6 +24,8 @@ inline uint16_t operator""_dbyte(unsigned long long val) {
 inline uint16_t operator""_dbytes(unsigned long long val) {
     return (uint16_t) val;
 }
+
+#endif
 
 } // namespace literals
 } // namespace data