+++ include/toolbox/strings/regex.h
@@ -12,7 +12,16 @@
 
 #include "toolbox/toolbox_config.h"
 
+#ifdef HAVE_REGEX_H
+
+#if defined(HAVE_STD_REGEX)
 #include <regex>
+namespace rxns = std;
+#elif defined(HAVE_BOOST_REGEX)
+#include <boost/regex.hpp>
+namespace rxns = boost;
+#endif //#if defined(HAVE_STD_REGEX)
+
 #include <string>
 #include <vector>
 
@@ -27,7 +36,7 @@ using match_all_res_t = std::vector<std::vector<std::string>>;
  * @param source
  * @return
  */
-TOOLBOX_API bool matches_pattern(const std::regex& pattern, const std::string& source);
+TOOLBOX_API bool matches_pattern(const rxns::regex& pattern, const std::string& source);
 /**
  * Match regex pattern.
  * @param pattern Just string. By default: search with flag
@@ -42,7 +51,7 @@ TOOLBOX_API bool matches_pattern(const std::string& pattern, const std::string& 
  * @param s
  * @return vector of vectors of strings
  */
-TOOLBOX_API match_all_res_t find_all_pattern(const std::regex& pattern, const std::string& source);
+TOOLBOX_API match_all_res_t find_all_pattern(const rxns::regex& pattern, const std::string& source);
 /**
  * Returns all found groups in source with pattern
  * @param pattern
@@ -56,7 +65,7 @@ TOOLBOX_API match_all_res_t find_all_pattern(const std::string& pattern, const s
  * @param source
  * @return
  */
-TOOLBOX_API std::string find_pattern_first(const std::regex& pattern, const std::string& source);
+TOOLBOX_API std::string find_pattern_first(const rxns::regex& pattern, const std::string& source);
 /**
  * Take first regex match and return it (with custom regex object)
  * @param pattern
@@ -74,7 +83,7 @@ TOOLBOX_API std::string find_pattern_first(const std::string& pattern, const std
  * @param source
  * @return
  */
-TOOLBOX_API std::vector<std::string> find_pattern(const std::regex& rxPattern, const std::string& source);
+TOOLBOX_API std::vector<std::string> find_pattern(const rxns::regex& rxPattern, const std::string& source);
 
 /**
  * Less boilerplate for std::regex
@@ -99,5 +108,7 @@ TOOLBOX_API bool num_is_real(const std::string& input);
 
 } // namespace strings
 } // namespace toolbox
+
+#endif //#ifdef HAVE_REGEX_H
 
 #endif // TOOLBOXPP_STRINGS_REGEX_H