+++ src/strings/regex.cpp
@@ -11,23 +11,25 @@
 
 #include <vector>
 
-bool toolbox::strings::matches_pattern(const std::regex& pattern, const std::string& source) {
-    std::smatch match;
-    return std::regex_search(source, match, pattern);
+#ifdef HAVE_REGEX_H
+
+bool toolbox::strings::matches_pattern(const rxns::regex& pattern, const std::string& source) {
+    rxns::smatch match;
+    return rxns::regex_search(source, match, pattern);
 }
 bool toolbox::strings::matches_pattern(const std::string& pattern, const std::string& source) {
-    return matches_pattern(std::regex(pattern, std::regex_constants::icase), source);
+    return matches_pattern(rxns::regex(pattern, rxns::regex_constants::icase), source);
 }
-std::vector<std::vector<std::string>> toolbox::strings::find_all_pattern(const std::regex& pattern,
+std::vector<std::vector<std::string>> toolbox::strings::find_all_pattern(const rxns::regex& pattern,
                                                                          const std::string& source) {
     std::vector<std::vector<std::string>> capturedGroups;
     std::vector<std::string> capturedSubgroups;
-    const std::sregex_token_iterator endIterator;
-    for (std::sregex_token_iterator it(source.cbegin(), source.cend(), pattern); it != endIterator; ++it) {
+    const rxns::sregex_token_iterator endIterator;
+    for (rxns::sregex_token_iterator it(source.cbegin(), source.cend(), pattern); it != endIterator; ++it) {
         capturedSubgroups.clear();
         std::string group = *it;
-        std::smatch res;
-        if (std::regex_search(group, res, pattern)) {
+        rxns::smatch res;
+        if (rxns::regex_search(group, res, pattern)) {
             for (size_t i = 0; i < res.size(); i++) { // NOLINT(modernize-loop-convert), sometimes foreach has
                 // strange effect - size is 0, but iterator have != 0 items
                 capturedSubgroups.push_back(res[i]);
@@ -43,12 +45,12 @@ std::vector<std::vector<std::string>> toolbox::strings::find_all_pattern(const s
 }
 std::vector<std::vector<std::string>> toolbox::strings::find_all_pattern(const std::string& pattern,
                                                                          const std::string& source) {
-    return find_all_pattern(std::regex(pattern, std::regex_constants::icase), source);
+    return find_all_pattern(rxns::regex(pattern, rxns::regex_constants::icase), source);
 }
-std::string toolbox::strings::find_pattern_first(const std::regex& pattern, const std::string& source) {
-    std::smatch results;
+std::string toolbox::strings::find_pattern_first(const rxns::regex& pattern, const std::string& source) {
+    rxns::smatch results;
     std::string result;
-    bool found = std::regex_search(source, results, pattern);
+    bool found = rxns::regex_search(source, results, pattern);
     if (!found || results.size() < 2) {
         return std::string();
     }
@@ -56,11 +58,11 @@ std::string toolbox::strings::find_pattern_first(const std::regex& pattern, cons
     return results[1];
 }
 std::string toolbox::strings::find_pattern_first(const std::string& pattern, const std::string& source) {
-    return find_pattern_first(std::regex(pattern, std::regex_constants::icase), source);
+    return find_pattern_first(rxns::regex(pattern, rxns::regex_constants::icase), source);
 }
-std::vector<std::string> toolbox::strings::find_pattern(const std::regex& rxPattern, const std::string& source) {
-    std::smatch result;
-    std::regex_search(source, result, rxPattern);
+std::vector<std::string> toolbox::strings::find_pattern(const rxns::regex& rxPattern, const std::string& source) {
+    rxns::smatch result;
+    rxns::regex_search(source, result, rxPattern);
 
     std::vector<std::string> out(result.size());
     const size_t cnt = result.size();
@@ -71,11 +73,13 @@ std::vector<std::string> toolbox::strings::find_pattern(const std::regex& rxPatt
     return out;
 }
 std::vector<std::string> toolbox::strings::find_pattern(const std::string& pattern, const std::string& source) {
-    return find_pattern(std::regex(pattern, std::regex_constants::icase), source);
+    return find_pattern(rxns::regex(pattern, rxns::regex_constants::icase), source);
 }
 bool toolbox::strings::num_is_integer(const std::string& input) {
-    return std::regex_match(input, std::regex(R"(^[-]?[0-9eE+]+?$)"));
+    return rxns::regex_match(input, rxns::regex(R"(^[-]?[0-9eE+]+?$)"));
 }
 bool toolbox::strings::num_is_real(const std::string& input) {
-    return std::regex_match(input, std::regex(R"(^[-+]?[0-9]*\.?[0-9]+([eE][-+]?[0-9]+)?$)"));
+    return rxns::regex_match(input, rxns::regex(R"(^[-+]?[0-9]*\.?[0-9]+([eE][-+]?[0-9]+)?$)"));
 }
+
+#endif //HAVE_REGEX_H